Fix !=0 check of divw, remw

+++ src/emulate.rs
@@ -837,7 +837,7 @@ fn exec_div(state: &mut EmulatorState, rtype: RType) {
 fn exec_divw(state: &mut EmulatorState, rtype: RType) {
     let rs1_value = state.get_reg(rtype.rs1());
     let rs2_value = state.get_reg(rtype.rs2());
-    assert!(rs2_value != 0, "check for non-zero divisor");
+    assert!((rs2_value as i32) != 0, "check for non-zero divisor");
     let rd_value = (rs1_value as i32).wrapping_div(rs2_value as i32) as u64;
     trace_rtype(state, "divw", rtype, rd_value);
     state.set_reg(rtype.rd(), rd_value);
@@ -873,7 +873,7 @@ fn exec_rem(state: &mut EmulatorState, rtype: RType) {
 fn exec_remw(state: &mut EmulatorState, rtype: RType) {
     let rs1_value = state.get_reg(rtype.rs1());
     let rs2_value = state.get_reg(rtype.rs2());
-    assert!(rs2_value != 0, "check for non-zero divisor");
+    assert!((rs2_value as i32) != 0, "check for non-zero divisor");
     let rd_value = (rs1_value  as i32).wrapping_rem(rs2_value  as i32)  as u64;
     trace_rtype(state, "remw", rtype, rd_value);
     state.set_reg(rtype.rd(), rd_value);
